app/config.py
# DATABASE_URL = "mysql+pymysql://root:root@localhost:3306/student_data_base"

# SECRET_KEY = "CHANGE_ME_TO_A_RANDOM_SECRET"
# ALGORITHM = "HS256"
# ACCESS_TOKEN_EXPIRE_MINUTES = 60


from dotenv import load_dotenv
import os

load_dotenv()  # Load the .env file

DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 60))

API_HOST = os.getenv("API_HOST", "127.0.0.1")
API_PORT = int(os.getenv("API_PORT", 8000))
